refactor: removed mongol_ prefix for mongol_helper
