+++ src/db/mongoldb/model/message.rs
@@ -4,7 +4,7 @@ use bson::Bson;
 use mongodb::bson::{DateTime, Uuid};
 use serde::{Deserialize, Serialize};
 
-use crate::{db::mongoldb::{as_string, mongol_helper}, model::{message::{Message, MessageFlag}, error}};
+use crate::{db::mongoldb::{as_string, helper}, model::{message::{Message, MessageFlag}, error}};
 
 
 #[derive(Debug, Serialize, Deserialize)]
@@ -28,15 +28,15 @@ impl TryFrom<&Message> for MongolMessage
 
     fn try_from(value: &Message) -> Result<Self, Self::Error>
     {     
-        let message_id = mongol_helper::convert_domain_id_to_mongol(&value.id)?;
+        let message_id = helper::convert_domain_id_to_mongol(&value.id)?;
         
-        let owner_id = mongol_helper::convert_domain_id_to_mongol(&value.owner.id)?;
+        let owner_id = helper::convert_domain_id_to_mongol(&value.owner.id)?;
         
-        let chat_id = mongol_helper::convert_domain_id_to_mongol(&value.chat.id)?;
+        let chat_id = helper::convert_domain_id_to_mongol(&value.chat.id)?;
 
         let bucket_id_option = value.bucket_id
             .as_ref()
-            .map(|bucket_id|mongol_helper::convert_domain_id_to_mongol(bucket_id))
+            .map(|bucket_id|helper::convert_domain_id_to_mongol(bucket_id))
             .transpose()?;
 
         let timestamp: SystemTime = value.timestamp.into();
